Add files via upload
 Understanding Partition Pruning in PySpark for Improved Query Performance
Introduction:
Partition pruning is a powerful optimization technique in PySpark that significantly enhances query performance by selectively reading and processing only the necessary partitions of data. In this tutorial, we'll explore the concept of partition pruning, its functionalities, and how it can be leveraged to boost the efficiency of your Spark applications.